requirements.txt
SpeechRecognition
PyAudio
pyttsx3
google-generativeai
python-dotenv
keyboard
pipwin # optional helper for installing PyAudio wheels on Windows 